backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
dnspython==2.4.2
email-validator==2.1.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
colorama==0.4.6
aiosmtplib==3.0.1
